MAJOR PATCH - Add Home Page, About Us page, and Grocery List CRUD (#44)
* fix: improved code to add/remove ingredients from grocery list

fixed the code to be more accurate to what the app is supposed to be. could still be improved, but it is better than before.

things changed:
- FoodItem class now has a name and a quantity
- instructions were updated
- ingredients are now displayed in one column
- can now add more than one of an ingredient using the +/- buttons

* feat: created home screen and navigation to it

a very simple home screen has been created. an icon has been added to the navigation bar, and when clicked, it will bring you to the home screen.

things changed:
- home screen was created with basic text
- added home icon to nav bar
- added home screen to the screens the app can navigate to

* fix: app opens to home screen

when the app is started, the home screen is now the first screen to be displayed

* fix: added more to home screen ui

added more elements to the home screen ui so that it looks more like our mockups. it still needs work, but it is a better representation of what the final product should look like.

things created:
- added placeholders for certain graphics
- added a button to navigate to the about us page (that currently navigates to the grocery list screen since the about us page doesn't exist yet)

* feat: created an about us page with proper navigation

a new about us page has been created. it is accessed by clicking a button on the home screen. it displays an unfinished mockup version of the final product

things created:
- a basic about us page with temporary information
- proper navigation to the about us page from the home screen