fix state for ingredients list
finally, after literal months, the ingredients' state is done properly.

lots of changes were made in this commit because it was very much either everything worked or nothing worked, so it was difficult actually figuring out when (good) progress was made.

either way, the IngredientsViewModel and IngredientsScreen files were updated so that the ingredients that show up on the ingredients screen are no longer hard coded and instead use StateFlow and Coroutines.

changes made:
- added an editable list in the viewModel to be able to update the individual ingredients.
- updated the increaseQuantity and decreaseQuantity functions to use coroutines and .update.
- created a copy method so that the state flow works properly in the viewModel.
- passed the viewModel into IngredientsScreen from MutsuApp.
- updated how the viewModel functions are being passed in IngredientsScreen.