Create new recipe screen (#65)
* feat: general layout of the screen

* fix: restructured code into more composables

* feat: first implementation of ingredients (displaying and adding)

Temporary solution until storage is implemented

* fix: added more functionality to ingredient input row

* fix: existing ingredients are now displayed

* implemented full recipe class in recipe information screen and refactored code to work with it

* feat: refactored recipe list screen

* FEAT: Navigation (#61)

* Update mockups in README.md (#51)

* Update README.md mockups

* Put images side by side

* Add placeholder

* Fix navigation (#52)

* Add screen destinations in MustuDestination

Created an interface and objects to represent the screens we are using so far.

* Set up NavController

To make sure the navigation is done properly, we needed to set up NavController.

Changes made:
- added navigation dependency to the project.
- created a new MutsuApp composable and added a NavHost to it.
- created a new folder for navigation files.

* Implemented navigation using NavController

set up a helper function to navigate to the different screens. These functions are now called in the onClick functions for each navbar element.

also updated recepe screen object name to be 'Recipes' instead of 'RecipeList'.

* access different screens from main app

moved the code that calls a specific screen from the MutsuDestination objects to the composables in the MutsuApp composable.

* access about us screen from home page

* remove old navigation code

* fix: Updated grocery list implementation and state (#59)

* display ingredients from flow and view model

the grocery list screen now uses a viewModel to get the list of ingredients to display. The state doesn't work yet, but everything is being displayed properly.

changes made:
- created an IngredientsViewModel.
- implemented the viewModel in the IngredientsScreen.
- changed the way ingredients are displayed so that they are set up in a foreach instead of individually.

* fix state for ingredients list

finally, after literal months, the ingredients' state is done properly.

lots of changes were made in this commit because it was very much either everything worked or nothing worked, so it was difficult actually figuring out when (good) progress was made.

either way, the IngredientsViewModel and IngredientsScreen files were updated so that the ingredients that show up on the ingredients screen are no longer hard coded and instead use StateFlow and Coroutines.

changes made:
- added an editable list in the viewModel to be able to update the individual ingredients.
- updated the increaseQuantity and decreaseQuantity functions to use coroutines and .update.
- created a copy method so that the state flow works properly in the viewModel.
- passed the viewModel into IngredientsScreen from MutsuApp.
- updated how the viewModel functions are being passed in IngredientsScreen.

* fix: can't remove ingredients that aren't on the list

made it so that the user can't have negative quantities of ingredients when clicking on the add/subtract buttons.

* remove old, unused code

---------

Co-authored-by: breannadefo <98484910+breannadefo@users.noreply.github.com>

* feat: first instance of RecipeViewModel and started navigation from RecipeList to RecipeInformation

* fix: made recipe list screen use view model

* feat: figuring out navigation from recipe list screen to recipe information screen

* feat: navigation from recipe list to recipe information with arguments

* feat: added functionality to save and remove buttons in recipe information screen

* fix: fixed big bug that caused recipe list screen to crash

* fix: fixed navigation from recipe list to recipe information screen (wasn't passing argument properly)

* fix: partially fixed textfield issue

* feat: fully fix textfield issue and add input validation

* feat: add first implementation of popup

---------

Co-authored-by: breannadefo <98484910+breannadefo@users.noreply.github.com>