add ability to delete an account
this ability is only possible when the user is logged in, and the only account they can delete is the one they are logged in as. There is also a pop up message that shows up when the delete button is clicked to make sure they want to delete their account.