FEAT: Implement Firebase for login/registration (#76)
* fix: rename things from EmptyActivity to Mutsu

not everything was changed because it wouldn't let me change everything, but there was a lot that changed. I ran the app after making these changes and it ran properly.

* update gradle to use firebase

* start setting up firebase logic

started adding the notes from Talib's slides into the project

changes made:
- created a new login/register folder and moved the LoginScreen code there.
- created a new class called Users. When the old Login code gets removed/revamped, I will rename this class to being User.
- created an interface for the authentication repository.

* add code to transfer current user information from firebase information

* implement the rest of the AuthRepository functions

all these functions exist in the AuthRepository interface and are being implemented in the AuthRepositoryFirebase class.

changes made:
- added signUp function.
- added signIn function.
- added signOut function.
- added delete function.

* create authentication view model

the view model has functions that call all the functions from the AuthRepository interface.

* add MyApp and AppModule files to be able to inject Firebase into our app

* create factory class for AuthViewModel

* fixed dependency for firebase

* add comments with credit of where I got most of the firebase code

* update login page to use firebase code

changed the code in the LoginRegisterScreen to use the information we need to login/register, as well as utilizing the AuthViewModel to make calls to Firebase.

changes made:
- added AuthViewModel to the parameters of the screen,
- initialized this viewmodel in MutsuApp.
- commented out previous code that is now outdated.
- used the viewmodel to get the current user.
- added textboxes and buttons to login/register.
- added a button to log out when the user is logged in.

* add ability to delete an account

this ability is only possible when the user is logged in, and the only account they can delete is the one they are logged in as. There is also a pop up message that shows up when the delete button is clicked to make sure they want to delete their account.

* add dispatchers to launch methods in AuthViewModel

a dispatchers value was passed into all the launch methods in the AuthViewModel which makes that code run on a different thread than the UI. This is supposed to make everything run smoother.

* change navigation to limit what is available for guests

now, the recipe screen, meal plan, and grocery list are only displayed and clickable on the navbar when the user is signed it. When they aren't signed in, they only have access to the home screen and the login/register screen.

* add basic input validation for login/registration

added a few checks of common errors people would make when entering their credentials to login/create an account. An error message pops up and the action doesn't go through.

some checks:
- both email and password have data in them.
- email follows the proper structure of an email.
- password is at least 8 characters long.

* remove old and unused code from login page

* feat: bring changes from main

* add google-services file that for some reason didn't add before

* made a few changes so that the app builds

changes made:
- removed problematic implementation.
- add import in food.kt.
- add import in main activity.
- change package name in proto file.

---------

Co-authored-by: Diana Karpeev <2059788@johnabbottcollege.net>