login/register user interface, confirmation popups and display meal plan ingredients list (#101)
* configure color theme and update home page UI

* feat: upgrade ingredients screen's user interface

* feat: improve user interface of recipe list and recipe information screens

* feat: add icon to scaffold top bar

* fix: slight color variations for accessibility

* fix: got started on meal plan screen

* fix: small correction to recipe information quantity column

* fix: small change to top bar's logo

* fix: change activity name to mutsu

* fix: tweak login register buttons

* feat: improve user interface and accessibility (#78) (#97)

* configure color theme and update home page UI

* feat: upgrade ingredients screen's user interface

* feat: improve user interface of recipe list and recipe information screens

* feat: add icon to scaffold top bar

* fix: slight color variations for accessibility

* fix: got started on meal plan screen

* fix: small correction to recipe information quantity column

* fix: small change to top bar's logo

* fix: change activity name to mutsu

* fix: refactor login/register forms

* feat: add toggle text under form

* feat: major improvement to UI of login-register

* feats: improve signed in screen ui and add password visual transformation

* start on accessibility for login register screen

* fix: fix login/registration error message bug and add documentation

* resolve merge conflict with main

* confirmation and display ingredients for grocery list (#100)

* add pop up that appears when the user tries to generate a grocery list

pop up shows up with a warning message and two options: dismiss or continue. So far only the dismiss part is properly implemented.

* show grocery list as a list

when the 'generate list' button is pressed and the popup is confirmed, the page switches to show a list of all the selected ingredients and their quantity

changes made:
- rememberSaveable boolean determines what is being shown on the screen.
- new composable function created to display the ingredients.
- fully implemented confirm part of popup.

* remove code that is no longer being used

* update instructions based on what's showing on screen

changed the instructions so that they tell the user what to do when they can edit the grocery list, but warn the user when they are just viewing the list. This warning is to let them know they will lose their list if they go to a different screen.

* fix ui of grocery list confirmation popup

changed the ui so the popup is more defined and easier to understand

* fix ui of generated grocery list

added a bit of padding and a background to make the grocery list look a little cleaner.

* change ui of alert for creating grocery list

modified it so that the background is white instead of beige and the border is gone

* change ui of alert for recipes

updated the ui of the alert dialogs used with the recipes to match the one used for creating a grocery list. Also added a reference to where we got the code for the alert dialogs

* fix icon display in alert dialogs

fixed the alert dialogs for when generating a grocery list and deleting your account so that their icon actually displays

* fix ui for alert when deleting your account

changed alert dialog ui to match the aler dialog ui used elsewhere in the app

---------

Co-authored-by: Breanna <bdeforest01@gmail.com>

* fix: adjust popup button's constrast

---------

Co-authored-by: Breanna <bdeforest01@gmail.com>