fix: update meal plan ui and usage of recipes from storage (#105)
* remove 'get started' button from home screen

this button wasn't doing anything since it was supposed to go to the login/register page, but that navigation is in the nav bar instead

* rename a few files to clearer names

changed the 'shopping list' file and composable to be 'meal plan' since that is what we have been referring this page at for a while.

* create meal display box

set up the box that will display all the meals. also commented out some code that isn't currently being used.

* create meals viewModel

create a viewModel to handle getting the recipes so that they can be displayed on the meal plan screen.

changes made:
- fixed recipes view model to seed certain recipes.
- created mealsViewModel.
- get recipes from storage in mealsViewModel.
- display name and quantity of meals on screen.
- pass mealsViewModel to meal screen from main activity.

* can increase meal quantity on meal plan screen

added an increase quantity function in the mealsViewModel and passed the function into the display box for the meal plan screen

* can decrease quantity of meal on meal plan screen

added a decrease function in the mealsViewModel and implemented it in the meal display box so that the quantity of a meal can decrease. Business logic doesn't allow the quanity to be a negative value.

* remove unused code from MealPlanScreen

fixed some spacing as well as removed old code

* add instructions to meal plan

added a bit of text at the top of the meal plan screen to give a brief explanation of what to do

* add button to go to grocery list from meal plan

the functionality isn't done, but the ui is