88 choose a license (#107)
* Create LISCENSE.md

Based on (currently 2/3 members) discussion, we think it best to go for the GNU public liscense.

* Update LISCENSE.md

* Delete LISCENSE.md

* Create LICENSE.md

I can't spell apparently, here's the real License file based on the conversations the three of us have now had. This will be reiterated when a PR is made, and the reviewer adds a comment signifying we agree with the license.