feat: firebase app distribution (#116)
* create distribution.yml

set up everything by using the video tutorial linked in the file and the teacher's example file

* update distribution.yml

adding on [push] to hopefully get this workflow to run

* add new on push triggers

add a trigger for when a push is made on this branch

* fix spelling mistake and versions

distribution.yml had a spelling mistake in one of the names. it also used version 1 for a couple actions so I changed it to using version 3 like in the tutorial

* enable debug to try and figure out what is going wrong

* add spaces to see if that was causing problems

whenever I run the workflow, I get a 400 invalid request error, so we updated the secrets again. hopefully this works because the only solution given by firebase support is to contact them with our information and they will check it out for us

* add release notes to firebase app distribution call

* insignificant change to trigger workflow

finally finished watching the tutorial and realized that our secrets were wrong. Fixed that so this should work now

* remove 'on push' since it works

we no longer need the 'on push' part since that was just for testing to make sure that everything ran properly